trader-edu-platform/requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
sqlalchemy==1.4.46
psycopg2-binary==2.9.5
transformers==4.26.0
tensorflow==2.11.0
numpy==1.24.2
pandas==1.5.3
websockets==10.4
pytest==7.2.1
redis==4.5.1
pydantic==1.10.4
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
aiohttp==3.8.4
matplotlib==3.7.0
scikit-learn==1.2.1 